feat/#4: 교수 전체 조회 기능 추가

+++ src/main/java/inha/git/user/api/controller/dto/response/CompanySignupResponse.java
@@ -10,4 +10,4 @@ public record CompanySignupResponse(
 
         @NotNull
         @Schema(description = "유저 아이디", example = "1")
-        Long userId){ }
+        Integer userId){ }
