+++ src/main/java/inha/git/user/api/controller/dto/response/StudentSignupResponse.java
@@ -10,4 +10,4 @@ public record StudentSignupResponse(
 
         @NotNull
         @Schema(description = "유저 아이디", example = "1")
-        Long userId){ }
+        Integer userId){ }
